FAR-12/refactor: 게시글 리팩토링
