--- src/main/java/com/ssafy/faraway/domain/post/controller/PostRestController.java
@@ -1,4 +1,0 @@
-package com.ssafy.faraway.domain.post.controller;
-
-public class PostRestController {
-}
