+++ src/main/java/com/ssafy/faraway/domain/post/repository/PostRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.faraway.domain.post.repository;
 import com.ssafy.faraway.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Long>, PostCustomRepository {
+public interface PostRepository extends JpaRepository<Post, Long> {
 }
